+++ NEWS.md
@@ -1,3 +1,5 @@
+# osrm.backend (development version)
+
 # osrm.backend 0.1.1
 
 * Fixes in test suite to address CRAN warnings.