+++ NEWS.md
@@ -1,4 +1,4 @@
-# osrm.backend (development version)
+# osrm.backend 0.2.0
 
 ## Breaking Changes